graphql with spring boot

+++ src/main/java/com/srv/app/annotation/service/impl/VechicleImpl.java
@@ -44,8 +44,14 @@ public class VechicleImpl implements VehicleService {
 
     @Override
     public Integer deletevehicle(Integer vehicleId) {
-        repo.deleteById(vehicleId);
-        return vehicleId;
+        if(checkvehicle(vehicleId)==Boolean.TRUE)
+        {
+
+            repo.deleteById(vehicleId);
+            return vehicleId;
+        }else {
+            return null;
+        }
     }
 
     @Override
